@@ -7,36 +7,38 @@ use Jumbo\Client\Options\ModuleOptions;
 
 class ModuleOptionsTest extends OptionsTestCase
 {
-    /**
-     * @var ModuleOptions
-     */
+    /** @var ModuleOptions */
     protected $options;
 
     protected function setUp()
     {
         $this->options = $this->getOptions(
             ModuleOptions::CLASS,
-            array(
+            [
                 'getClient',
                 'getClients',
                 'setClients',
-            )
+            ]
         );
     }
 
-    public function testOptionsExist()
+    public function testOptionsExist(): void
     {
         $this->assertInstanceOf(ModuleOptions::CLASS, $this->options);
     }
 
-    public function testClientCanBeSet()
+    public function testClientCanBeSet(): void
     {
+        $clientConfig = [
+            'base_uri' => 'http://some.client.com',
+            'app_id' => 'Some-App-ID',
+            'app_key' => 'Some-App-Key',
+        ];
+
         $clientName = 'SomeClient';
-        $clientsConfig = array(
-            $clientName => array(
-                'base_url' => 'http://some.client.com',
-            ),
-        );
+        $clientsConfig = [
+            $clientName => $clientConfig,
+        ];
 
         $this->assertTrue(is_array($this->options->getClients()));
         $this->assertCount(0, $this->options->getClients());
@@ -45,6 +47,13 @@ class ModuleOptionsTest extends OptionsTestCase
         $this->options->setClients($clientsConfig);
 
         $this->assertCount(1, $this->options->getClients());
-        $this->assertInstanceOf(ClientOptions::CLASS, $this->options->getClient($clientName));
+
+        $clientOptions = $this->options->getClient($clientName);
+
+        $this->assertInstanceOf(ClientOptions::CLASS, $clientOptions);
+        $this->assertEquals($clientConfig['base_uri'], $clientOptions->getBaseUri());
+        $this->assertEquals($clientConfig['app_id'], $clientOptions->getAppId());
+        $this->assertEquals($clientConfig['app_key'], $clientOptions->getAppKey());
+        $this->assertEquals($clientConfig, $clientOptions->toArray());
     }
 }
